Seperates each table into .py files
